README.md 수정 및 패키지 일부 보완
