+++ edge-service/src/main/java/com/koreacb/msa/edge/client/LoanClient.java
@@ -2,7 +2,6 @@ package com.koreacb.msa.edge.client;
 
 import org.springframework.cloud.openfeign.FeignClient;
 
-@FeignClient(name = "card")
-public interface CardClient {
-
+@FeignClient(name = "loan")
+public interface LoanClient {
 }
