+++ edge-service/src/main/resources/application.yaml
@@ -1,5 +1,5 @@
 server:
-  port: 14000
+  port: 18080
   servlet:
     context-path: /
